requirements.txt
beautifulsoup4
requests
nltk
pyfood
ingredient-parser-nlp